1D Thrust + More CMUG
Thrust is now functionally one-directional, as in, pushes in the direction the player is rotated rather than being controlled componentwise. Also ported cmugraphics into the subfolder for testbeds, cause I don't want to fool around with 'dependent on having cmu_graphjics in the same folder'. So pardon the file change size