Collisions and Cleanup
Collisions are successfully implemented between smallbodies and sphereplanets! Deliciously, if I might add, with rubber-banding prevention and variable elasticity to boot. I have not tested it on a non-player small body, but all the code is dragged from smallbody superclass to player anyway, so it should have no trouble.

Line collisions, gravity-aligned camera, and so much more, here we come!!!!