Intro to Images and Screens
Tinkered with images and screens. Got sprite reference working (as per E's advice: find the file in question in the VSCode workspace dropdown to the left, simply ctrl+c it, and paste that within quotes into the code to get its relative address, which can be directly cited by cmugraphics' image tool. Also worked several screens into the mix: it provides an interesting quandary of when I would ever truly want my simulation to not be running (During start, most certainly? But then again, it would be wonderful to simply have GUI elements float on the screen while the sim IS 'active' and simply disable player inputs and clear the text once we begin)

I digress. I got images working!