Render Class
Did some restructuring. Render code is now fully internal to a dedicated class, and draw calls are made with render.getCoords, .getAngle, and .getScale. Very nice. I can't fully divorce it from app variables - it depends on knowing app width and height, curse this software - but it was a noble effort to divorce it from the app stuff.