LINEPLANET COLLISION REACHED!
Line planet collision is perfectly working! I could not be happier. I don't have a way to display planet mesh and not collision mesh, they are one and the same in this situation, and so I want to come up with a computational solution (which also will allow each differently sized smallbody to evaluate to a working custom collision mesh).