LINEAR MOVING PLANET COLLISIONS
With the addition of a new Motion library and some subtraction math inside collision evaluation, smallbodies now behave properly when colliding with (linearly) moving planets! Internal to collision evaluation are two functions that convert from global to planet-relative velocity and back, allowing collision math to be run in the planet's inertial frame. What a wondrous day.