Images + Failure Checkpoint
Lots of images have been made for the player sprite. Unfortunately, however, transparency does not register on them. I am tinkering with that, as well as with the motion pathing for planets, neither are in a complete state. I will not leave them off for the night - ideally, setting a safe time of around two in the morning, I will have these mysteries solved. (PIL offers a way to make these images functionally transparent, and the pathing just requires more thought.)

Also, does the ship need to have its main body not be a shade of orange / brown? I am starting to find it too homogenous. I may make it a big purple barrel, or perhaps a gray / lilac adjacent shade.

Exciting!