Rotation paths success!
And there was much rejoicing! My handler for rotation paths is working wonderfully. It has to work through some fascinating kinks for rotation. For exapmle, how any colliding smallbody must immediately start corotating with the planet it's on, (which is accounted for with a direct rotation modification), as well as a couple layers of initalization functions that spread out the load of defining these things. In fact, actually, smallbody collision recieved an entire new function dedicated to corotating with a colliding planet, which is actually really fun to see on its own if you disable gravity and collision-eval and just experience being 'trapped' by a linebody and watching it bring you perfectly to its motion frame.

This, all in all, has also been a very good foothold in the way to orbit paths. I designed those such that they may only need to account for linear movement; so the majority of my rotation-related headaches end now. Then again, those may be the sleep deprivation headaches. Nonetheless, they will disappear soon. I'm not exactly where I want to be, but that can be figured out in another day, and no matter what, I've got a gorgeous project already.