Transparent sprites, first planets, velocity indicator
The time approaches! I am getting concerned that a few things are.... unlikely to make it into this game. However, most things work now!

- Transparent sprites
- Somber heart and bramble triplets
- Diagetic landing mode
- BEAUTIFUL velocity indicator

Things still to do
- fruit :(
- probably simple fruit spawners
- player fruit collection in some manner
- score system for depositing the fruit in a bucket on the home planet

- player gravity moves downwards to bottom of screen (dear god this one will haunt me)
- other meaningful indicator of where gravitational bodies are

OOOH and a parallax star background. That seems like it might be a doozy. Do we suppose my strategy still, giving it 99% of the player's (linear) velocity in worldspace, will work? I should try it.

Oh well. How fun it all is, though! And i suppose tomorrow, I get to bring it together artistically.