BLACK HOLES, START SCREENS, HOME
Great scott, what a beautiful introductory page!
- Start and pause menus with beautiful variable-opacity texxt
- Functional black hole now at the core of the world, made a class that can use this anywhere
- Reset button to instantly teleport back to the starting position on Lazuna Vine

All that remains is, really, the fruit. I've put down some more logic in the Motion library for spawning things relative to the surface of circular bodies. Black hole uses it to teleport objects outside of the corresponding white hole. I can use it similarly to have spawners not maintain their own position, but just spawn a fruit in when the time is right.

I hope that the combination of logic I've developed over the instance-culling through screenhandler and the circular movement via black holes will finally cumulate in getting this damn thing up and running. After that, just a bit of art and video!