FRUIT LAUNCHING IS A GO
Fruits spawn consistently! Player can pick them up! And they are then appended to an internal list where adedicated class object, the player fruit launcher, can send them out the front of the player! Oh glorious day, fruit collection is a working mechanism in spacefruit as of right now.

All that's left is one more class object, with some traits of inventory appending much like the way the player does it. If we can instantiate it to just hang out on the bottom of Lazula Vine, where we can kinda encloes it in a lineplanet to represent the Big Bucket, then we can have our PHYSICAL score barrel!!! How cool! And it would, I suppose, keep a runnning tally of both score AND how mnay fruits have been delivered to the bucket (which might be nice to view from the pause menu, or a nearby sign)

All in all though, all of that after today's lecture, methinks.

Oh dear. That leaves nary the afternoon for... planets, art, AND the video. Oof.